+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.6] - unreleased
+
+Output is now colour-coded by default, override with `--color/--no-color`
+or by setting `WHATNEXT_COLOR`.
+
+
 ## [v0.5] - 2025-12-01
 
 Tasks can have deadlines using the `@YYYY-MM-DD/4w` syntax. As the deadline